pyproject.toml
[project]
name = "vecna"
version = "0.1.0"
description = "A terminal AI coding assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.76.0",
    "ipdb>=0.13.13",
    "ipython>=9.9.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
vecna = "vecna.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vecna"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "UP",  # pyupgrade (modern Python syntax)
]

[tool.ruff.lint.isort]
known-first-party = ["vecna"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
